% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRandomNumbers.R
\name{sampleRandomNumbers}
\alias{sampleRandomNumbers}
\title{Illustration of the random number generator in R. We focus here on the normal distribution}
\usage{
sampleRandomNumbers(sampleSize = 100)
}
\arguments{
\item{sampleSize}{size of the sample}
}
\value{
a list containing: the empirical mean of sample 1, the empirical variance of sample 1, the empirical variance of sample 2, the empirical variance of sample 3
}
\description{
This function first generates a sample of size sampleSize, drawn from a Normal distribution 
with mean 0 and variance 1
This sample is called Sample 1. Next, using observations from Sample 1, it generates Sample 2, 
containing observations that follows a normal distribution with mean 6 and variance 1.
Last, still using observations from Sample 1, it generates Sample 3, 
containing observations that follows a normal distribution with mean 6 and variance 4.
}
\examples{
U<-sampleRandomNumbers(1000)

U[[1]] # the Empirical Mean of Sample 1
U[[2]] # the Empirical Variance of Sample 1
U[[3]] # the Empirical Variance of Sample 2
U[[4]] # the Empirical Variance of Sample 3
}
